% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{Barplot for binomial distribution and probabilities}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{number of iterations}

\item{n}{sample size}

\item{p}{probability of a success}
}
\value{
plot of binomial distribution and random variable probabilities
}
\description{
Barplot for binomial distribution and probabilities
}
\examples{
mybin(10000, 10, 0.3)
}
