% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Perform bootstrap simulation technique for a sample}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{number of iterations to sample for}

\item{x}{sample to resample from}

\item{fun}{function to apply to the samples}

\item{alpha}{value to define confidence interval}

\item{cx}{character size}

\item{...}{additional parameters}
}
\value{
List containing ci, fun, x, and xstat
}
\description{
Perform bootstrap simulation technique for a sample
}
\examples{
myboot2(x=c(1,1,2,7,4))
}
