% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltb.R
\name{mycltb}
\alias{mycltb}
\title{Histogram of sample means for binomial distribution}
\usage{
mycltb(n, iter, p = 0.5, ...)
}
\arguments{
\item{n}{sample size}

\item{iter}{number of samples/iterations}

\item{p}{probability of success}

\item{...}{additional arguments}
}
\value{
Histogram and theoretical curve of the sample means
}
\description{
Histogram of sample means for binomial distribution
}
\examples{
mycltb(n=20,iter=10000,p=0.3)
}
