% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{Generates plot for maximum likelihood}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{Log likelihood function}

\item{x}{Data values}

\item{param}{Range of parameter to estimate}

\item{...}{Additional arguments}
}
\value{
List containing index of max, estimated value, y-value, and slopes
}
\description{
Generates plot for maximum likelihood
}
\examples{
logbin=function(x,param) log(dbinom(x,prob=param,size=10))
mymaxlik(x=c(9,9,1,9,9,9),param=seq(0,1,length=1000),
lfun=logbin,xlab=expression(p),main="Binomial",cex.main=2)
}
