% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Plot for normal distribution and area of P(X<=a)}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{mean}

\item{sigma}{standard deviation}

\item{a}{value to calculate P(X<=a)}
}
\value{
P(X<=a) in a list
}
\description{
Plot for normal distribution and area of P(X<=a)
}
\examples{
myncurve(5, 10, 3)
}
