% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Find the optimal number of tickets to sell for a flight}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{number of seats on flight}

\item{gamma}{probability of overbooking}

\item{p}{probability of a person showing}
}
\value{
list containing optimal tickets for discrete and continuous, along with parameters
}
\description{
Find the optimal number of tickets to sell for a flight
}
\examples{
ntickets(N=400, gamma=0.02, p=0.95)
}
